# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Epic Poetry Cafe Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales entries
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesEntry"
    post:
      operationId: createSalesEntry
      tags: [sales]
      summary: Create sales entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesEntry"
  /sales/{id}:
    patch:
      operationId: updateSalesEntry
      tags: [sales]
      summary: Update sales entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesEntryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesEntry"
    delete:
      operationId: deleteSalesEntry
      tags: [sales]
      summary: Delete sales entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sales/daily-summary:
    get:
      operationId: getDailySalesSummary
      tags: [sales]
      summary: Daily sales summary
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySalesSummary"
...
  /trials:
    get:
      operationId: listTrials
      tags: [trials]
      summary: List trials
      responses:
        "200":
          description: Trials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trial"
    post:
      operationId: createTrial
      tags: [trials]
      summary: Create trial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrialRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trial"
  /trials/{id}:
    get:
      operationId: getTrial
      tags: [trials]
      summary: Get trial detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trial detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialDetail"
    patch:
      operationId: updateTrial
      tags: [trials]
      summary: Update trial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrialRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trial"
  /trials/{id}/versions:
    post:
      operationId: createTrialVersion
      tags: [trials]
      summary: Create trial version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrialVersionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialVersion"
  /trials/{trialId}/versions/{versionId}/convert:
    post:
      operationId: convertTrialToMenuItem
      tags: [trials]
      summary: Convert approved trial to menu item recipe
      parameters:
        - name: trialId
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Converted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
...
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report as CSV
      parameters:
        - name: reportType
          in: query
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        sellingPrice:
          type: number
        dineInPrice:
          type: ["number", "null"]
        takeawayPrice:
          type: ["number", "null"]
        deliveryPrice:
          type: ["number", "null"]
        productionCost:
          type: number
        margin:
          type: number
        marginPercent:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - sellingPrice
        - productionCost
        - margin
        - marginPercent
        - active
...
    SalesEntry:
      type: object
      properties:
        id:
          type: integer
        salesDate:
          type: string
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: number
        sellingPrice:
          type: number
        totalAmount:
          type: number
        discount:
          type: number
        channel:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salesDate
        - menuItemId
        - menuItemName
        - quantity
        - sellingPrice
        - totalAmount
        - channel
    CreateSalesEntryRequest:
      type: object
      properties:
        salesDate:
          type: string
        menuItemId:
          type: integer
        quantity:
          type: number
        sellingPrice:
          type: number
        discount:
          type: number
        channel:
          type: string
        notes:
          type: string
      required:
        - salesDate
        - menuItemId
        - quantity
        - sellingPrice
        - channel
    DailySalesSummary:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        totalQuantity:
          type: number
        itemCount:
          type: integer
        topItems:
          type: array
          items:
            $ref: "#/components/schemas/TopSellingItem"
      required:
        - date
        - totalSales
        - totalQuantity
        - itemCount
        - topItems
    TopSellingItem:
      type: object
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: number
        revenue:
          type: number
      required:
        - menuItemId
        - menuItemName
        - quantity
        - revenue
...
    Trial:
      type: object
      properties:
        id:
          type: integer
        trialCode:
          type: string
        proposedItemName:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        targetCost:
          type: ["number", "null"]
        targetSellingPrice:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - trialCode
        - proposedItemName
        - status
    TrialDetail:
      type: object
      properties:
        trial:
          $ref: "#/components/schemas/Trial"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/TrialVersion"
      required:
        - trial
        - versions
    CreateTrialRequest:
      type: object
      properties:
        proposedItemName:
          type: string
        categoryId:
          type: integer
        targetCost:
          type: number
        targetSellingPrice:
          type: number
        notes:
          type: string
      required:
        - proposedItemName
    UpdateTrialRequest:
      type: object
      properties:
        proposedItemName:
          type: string
        categoryId:
          type: integer
        targetCost:
          type: number
        targetSellingPrice:
          type: number
        status:
          type: string
        notes:
          type: string
    TrialVersion:
      type: object
      properties:
        id:
          type: integer
        trialId:
          type: integer
        versionNumber:
          type: integer
        batchSize:
          type: number
        yieldQty:
          type: number
        yieldUom:
          type: string
        prepTime:
          type: ["integer", "null"]
        totalCost:
          type: number
        costPerUnit:
          type: number
        status:
          type: string
        tasteScore:
          type: ["number", "null"]
        appearanceScore:
          type: ["number", "null"]
        consistencyScore:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/TrialIngredientLine"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - trialId
        - versionNumber
        - batchSize
        - yieldQty
        - yieldUom
        - totalCost
        - costPerUnit
        - status
        - ingredients
    TrialIngredientLine:
      type: object
      properties:
        ingredientId:
          type: integer
        ingredientName:
          type: string
        plannedQty:
          type: number
        actualQty:
          type: number
        uom:
          type: string
        wastageQty:
          type: number
        costPerUnit:
          type: number
        totalCost:
          type: number
      required:
        - ingredientId
        - ingredientName
        - plannedQty
        - actualQty
        - uom
        - totalCost
    CreateTrialVersionRequest:
      type: object
      properties:
        batchSize:
          type: number
        yieldQty:
          type: number
        yieldUom:
          type: string
        prepTime:
          type: integer
        status:
          type: string
        tasteScore:
          type: number
        appearanceScore:
          type: number
        consistencyScore:
          type: number
        notes:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/TrialIngredientInput"
      required:
        - batchSize
        - yieldQty
        - yieldUom
        - ingredients
    TrialIngredientInput:
      type: object
      properties:
        ingredientId:
          type: integer
        plannedQty:
          type: number
        actualQty:
          type: number
        uom:
          type: string
        wastageQty:
          type: number
      required:
        - ingredientId
        - plannedQty
        - actualQty
        - uom
...